+++ client/game.css
@@ -1,44 +1,6 @@
 /*웹 폰트 사용*/
 @import url('https://fonts.googleapis.com/css2?family=Black+Han+Sans&display=swap');
 
-
-#loginTitle {
-    margin: 0;
-    margin-bottom: 20px;
-    width: 100%;
-    padding: 0;
-    font-family: 'Black Han Sans', sans-serif; /*웹폰트*/
-    font-size: 1.99rem;
-    text-align: center;
-    color:whitesmoke;
-}
-
-#signDiv { /*로그인창*/
-    display:none;
-    text-align: center;
-    margin-top: 50px;
-    width: 300px;
-    padding: 30px;
-    border: none;
-    box-shadow: 0 0 30px 3px rgba(0, 0, 255, 0.3);
-    background-color: rgba(0, 0, 205, 0.3);
-    /*background-image: url(sprites/loginBack.png);*/
-    background-size: 100% 100%;
-    color:whitesmoke;
-
-    border-radius: 10px; /*테두리를 둥글게*/
-
-    position: fixed;
-    top: 30%;
-    left: 50%;
-    margin-top: -100px;
-    margin-left: -200px;
-
-    font-family: 'Black Han Sans', sans-serif; /*웹폰트*/
-
-    animation:fadein 3s; /*서서히 나타나는(fade in) 애니메이션 효과 적용*/
-}
-
 @keyframes fadein{/*서서히 나타나는(fade in) 애니메이션 효과*/
   from{
     opacity:0;
@@ -48,59 +10,11 @@
   }
 }
 
-
-#signDiv input{ /*로그인창 내부의 닉네임 입력란*/
-    border: 0;
-    
-    /*background: none;*/
-    background-color: rgba(0,0,0,0.6);
-    display: block;
-    margin: 10px auto;
-    text-align: center;
-    width: 150px;
-    height: 35px;
-    border: 1px solid #45a29e;
-    padding: 5px;
-    color: #5c6c7c;
-    outline: none;
-    border-radius: 10px;
-    transition: 0.5s;
-    cursor: pointer;
-}
-#signDiv input[type ="text"]:focus{ /*로그인창 내부의 닉네임 입력란 클릭 효과*/
-    width: 250px;
-    border-color: #fce181;
-    border-radius: 20px;
-  } 
-  
-.button {
-    background-color: #4CAF50; /* Green */
-    border: none;
-    color: white;
-    padding: 15px 32px;
-    text-align: center;
-    text-decoration: none;
-    display: inline-block;
-    font-size: 13px;
-    margin: 4px 2px;
-    border-radius: 9px;
-}
-.button:active{
-    background-color: #234525; /* Green */
-    border: none;
-    color: white;
-    padding: 15px 32px;
-    text-align: center;
-    text-decoration: none;
-    display: inline-block;
-    font-size: 12px;
-    margin: 4px 2px;
-    border-radius: 8px;
-}
 /*
 게임화면 canvas
 주의 : css에서 크기를 확대/축소 하면, 그냥 비율이 확대된다.
 */
+
 #myCanvas{ 
     width:100%;
     height:100%;
@@ -109,50 +23,9 @@
     background-color: #2e8b57;
 }
 
-#background {
-    zoom: 150%;
-}
+/*자동 확장기능 overflow-y: auto; */
 
-.char-select-button {
-    background-color: white;
-    border: none;
-    text-align: center;
-    font-size: 12px;
-    font-weight: bold;
-}
-/*캐릭터 선택 버튼 효과*/
-.char-select-button:hover {
-    background-color: gray;
-    border: none;
-    text-align: center;
-    font-size: 12px;
-    font-weight: bold;
-}
-
-#chat-text{
-    width: 500px;
-    height : 100px; 
-    overflow-y: auto;
-}
-
-#character-select{
-    float: left;
-    padding: 25px;
-}
-
-#chat {
-    float:left;
-}
-
-#extrabuttons{
-    padding-left: 50px;
-}
-
-/*
-jeho edited
-UI 디자인
-*/
-.player-list-box{
+#ui_player_list_box{
     width: 15vw;
     height: 80vh;
     padding: 20px;
@@ -166,7 +39,7 @@ UI 디자인
     animation:fadein 3s;
   }
 
-  /*----동적생성 UI -----*/
+  /*----로그인 UI창 -----*/
 
 .ui-block{/*공중에 떠있는 투명한 검은색 ui 틀 // 체력 등 따로 관리할 것들*/
     z-index: 1000; /*쌓임높이. 높을수록 겉에 표시됨*/
@@ -174,63 +47,87 @@ UI 디자인
     border-radius: 30px;
 }
 
-.ui{ /*// sign-div 복사해놓은것, 화면의 공통적인 요소*/
+.ui{ /*로그인 ui 전체*/
     text-align: center;
-    /*margin-top: 50px;*/
     padding: 30px;
     border: none;
     box-shadow: 0 0 30px 3px rgba(0, 0, 255, 0.3);
     background-color: rgba(0, 0, 205, 0.3);
-    /*background-image: url(sprites/loginBack.png);*/
     background-size: 100% 100%;
     color:whitesmoke;
-
     border-radius: 10px; /*테두리를 둥글게*/
-
     position: fixed;
-    /*
-    top: 30%;
-    left: 50%;
-    margin-top: -100px;
-    margin-left: -200px;
-    */
+    animation:fadein 3s;
 }
 
 .title{
     width: 50vw;
     height: 20vh;
     top: 10%;
-    left: 5%;
+    left: 20%;
     font-family: 'Black Han Sans', sans-serif; /*웹폰트*/
-    font-size: 1.99rem;
+    font-size: 4rem;
     background-color: rgba(0, 0, 0, 0);
     box-shadow: 0 0 30rem 3rem rgba(0, 0, 0, 0);
     color: #fffff0;
-    animation:fadein 2s;
 }
 
-.username_input{
+#username_input{
     width: 15vw;
     height:7vh;
     top: 35%;
-    left: 20%;
+    left: 22%;
     padding: 28px;
     background-color: rgba(0, 0, 0, 0.4);
+    transition:0.5s;
 }
 
-.access_button{
+#username_input:focus{
+    width: 18vw;
+    border-color: #fce181;
+    border-radius: 20px;
+    transition:0.5s; /*서서히 커짐*/
+}
+
+#play_button {
     width: 10vw;
     height: 8vh;
-    top: 52%;
-    left: 23.5%;
-    padding: 3px;
-    font-family: 'Black Han Sans', sans-serif; /*웹폰트*/
-    font-size: 1.2vmax;
-    background-color: rgba(0, 0, 0, 0);
-    box-shadow: 0 0 30px 3px rgba(0, 0, 0, 0);
-    color: #fffff0;
+    top: 45%;
+    left: 25%;
+    font-size: 25px;
+    padding: 15px 30px;
     border: 3px solid #fffff0;
+    background-color: transparent;
+    color: #fffff0;
+    text-transform: uppercase;
+    letter-spacing: 5px;
+    font-weight: bold;
+    transition: all 0.4s;
+    overflow: hidden;
 }
+#play_button:focus {
+    outline: none;
+}
+
+#play_button::before {
+    content: "";
+    position: absolute;
+    height: 100%;
+    width: 100%;
+    background-color: #fffff0;
+    top: 100%;
+    left: 0;
+    transition: all 0.4s;
+    z-index: -1;
+}
+
+#play_button:hover {
+    color: #3c3e70;
+}
+#play_button:hover::before{
+    transform:translateY(-100%);
+}
+
 
 .charactor-select{
     width: 25%;
@@ -241,23 +138,28 @@ UI 디자인
 }
 
 .how-to-play{
-    top: 68%;
+    top: 55%;
     left: 27%;
     padding: 3px;
     background-color: rgba(0, 0, 0, 0);
     box-shadow: 0 0 30px 3px rgba(0, 0, 0, 0);
     color: #fffff0;
-    font-size: 0.6vmin;
+    font-size: 1.5vmin;
+    transition:0.3s;
+}
+.how-to-play:hover{
+    font-size: 2vmin;
+    transition:0.3s;
 }
 
 .guide{
-    top: 73%;
-    left: 23%;
+    top: 58%;
+    left: 27%;
     padding: 3px;
     background-color: rgb(117, 109, 109, 0.5);
     box-shadow: 0 0 30px 3px rgba(0, 0, 0, 0);
     color: #fffff0;
-    animation:fadein 3s;
+    transition:0.3s;
 }
 
 input::placeholder {
