+++ client/scripts/ui.js
@@ -1,0 +1,68 @@
+//
+// Ui.js
+//
+Ui={
+
+    create_login_ui:function(){
+        let ui_login = document.getElementById("ui_login");
+        let ui_game = document.getElementById("ui_game");
+
+        //동적 ui 생성
+        const ui_title = document.createElement('div');
+        ui_title.classList.add('ui');
+        ui_title.classList.add('title');
+        ui_title.innerHTML='Magic Blocker';
+
+        ui_login.appendChild(ui_title);
+
+        const ui_name_input = document.createElement('input');
+        ui_name_input.classList.add('ui');
+        ui_name_input.classList.add('username_input');
+        ui_name_input.setAttribute('placeholder','Please Enter Nickname');
+
+        ui_login.appendChild(ui_name_input);
+
+        const ui_play_button = document.createElement('button');
+        ui_play_button.classList.add('ui');
+        ui_play_button.classList.add('play_button');
+        ui_play_button.innerHTML='Play';
+        ui_login.appendChild(ui_play_button);
+
+        const ui_charactor_select = document.createElement('div');
+        ui_charactor_select.classList.add('ui');
+        ui_charactor_select.classList.add('charactor-select');
+        ui_charactor_select.innerHTML="여기에 캐릭터 선택창 떠야됨"
+
+        ui_login.appendChild(ui_charactor_select);
+
+        const ui_how_to_play_button = document.createElement('button');
+        ui_how_to_play_button.classList.add('ui');
+        ui_how_to_play_button.classList.add('how-to-play');
+        ui_how_to_play_button.innerHTML='How to Play?';
+
+        ui_login.appendChild(ui_how_to_play_button);
+
+        const ui_guide_page = document.createElement('div');
+        ui_guide_page.classList.add('ui');
+        ui_guide_page.classList.add('guide');
+        ui_guide_page.id = "guideID";
+        ui_guide_page.innerHTML='이동 : w a s d \n 발사 : k';
+
+        const ui_player_list_box = document.createElement('div');
+        ui_player_list_box.id = 'ui_player_list_box';
+        ui_player_list_box.innerHTML='접속중인 플레이어';
+        ui_game.appendChild(ui_how_to_play_button);
+
+        ui_how_to_play_button.onclick = function(){ // How to Play? 클릭 시 도움말 뜨도록
+            if(document.getElementById("guideID")==null){
+                ui_login.appendChild(ui_guide_page);
+            }else{
+                ui_guide_page.remove();
+            }
+            
+        }
+
+    }
+}
+
+
