--- prTest.txt
@@ -1,2 +1,0 @@
-pull request test..
-merge test...
